backend/requirements.txt
flask
flask-cors
pynguin
openai>=1.0.0
python-dotenv>=0.19.0
